@@ -1,13 +1,12 @@
 {
-    "private": true,
-    "name": "homey-switcher",
+    "name": "il.co.switcher",
     "version": "1.0.0",
     "main": "app.ts",
     "devDependencies": {
-        "@tsconfig/node12": "^1.0.11",
-        "@types/homey": "npm:homey-apps-sdk-v3-types@^0.3.4",
-        "@types/node": "^18.16.16",
-        "typescript": "^4.9.5"
+        "@tsconfig/node16": "^16.1.1",
+        "@types/homey": "npm:homey-apps-sdk-v3-types@^0.3.5",
+        "@types/node": "^20.10.6",
+        "typescript": "^5.3.3"
     },
     "dependencies": {
         "switcher-js2": "^1.5.5"
