@@ -1,3 +1,3 @@
-import SwitcherDevice from "../Device";
+import SwitcherDevice from '../Device';
 
 module.exports = class SwitcherMiniDevice extends SwitcherDevice {}
